git-svn-id: http://ctp2.darkdust.net/anonsvn/branches/linux-rework@675 34d2fc0c-33df-0310-8a34-85f8040ba873

+++ ctp2_code/os/include/ctp2_inttypes.h
@@ -19,8 +19,8 @@ typedef signed char          sint8;
 typedef unsigned char        uint8;
 typedef signed short int     sint16;
 typedef unsigned short int   uint16;
-typedef signed int           sint32;
-typedef unsigned int         uint32;
+typedef signed long int      sint32;
+typedef unsigned long int    uint32;
 typedef signed __int64       sint64;
 typedef unsigned __int64     uint64;
